package/gumbo-parser: bump to version 0.13.1
Changelog:
https://codeberg.org/gumbo-parser/gumbo-parser/releases

Release 0.11.0 onwards dropped the 'v' version prefix in the tarball
name, so drop it on our side as well.

The patch no longer applied cleanly and needed to be refreshed.
It was also submitted to the new upstream, the upstream link is
updated accordingly.
A local test build was done with the toolchain from the config that
originally triggered the build to fail due to lack of C++ support:
http://autobuild.buildroot.org/results/a32b5d3b959433fd5c3543661c37f80d27fbd010

It is worth noting that we still build using autotools which is
deprecated in this version of gumbo-parser (in favor of meson).

The license file location was moved in a "doc" subdirectory, in upstream
commit [1]. This commit reflect that change.

[1] https://codeberg.org/gumbo-parser/gumbo-parser/commit/814e2b43f776e25ffe2089396ced63e6ceb4db06

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>
[Julien: fix license file location which has moved to "doc/"]
Signed-off-by: Julien Olivain <ju.o@free.fr>