toolchain/toolchain-external/toolchain-external-custom: hide gcc versions for unsupported CPUs
The BR2_ARCH_NEEDS_GCC_AT_LEAST_X guards were added in commit [1]
for toolchain-external-custom gcc versions from 4.8 to 12. No guard
was added for gcc 13 because is was the latest version at that time
(and the symbol BR2_ARCH_NEEDS_GCC_AT_LEAST_14 was not available
at that time).

Then, commit [2] and [3] added gcc 14 and 15 without adding
guards.

In order to prevent a Buildroot user selecting an external custom
toolchain that does not support a selected CPU, this commit adds all
the missing guards.

[1] https://gitlab.com/buildroot.org/buildroot/-/commit/eed1670d8a6f0899f5fdfc26bd6d1609ebc19c7a
[2] https://gitlab.com/buildroot.org/buildroot/-/commit/7fa12e4f9ed736c7238dbdadf4dc5b9da03007b3
[3] https://gitlab.com/buildroot.org/buildroot/-/commit/5f1a38a41f0b6702444940b24bcd0d79a41a0059

Reported-by: Arnout Vandecappelle <arnout@rnout.be>
Signed-off-by: Julien Olivain <ju.o@free.fr>
Signed-off-by: Romain Naour <romain.naour@smile.fr>