package/petitboot: add patches to fix build errors with "opaque" ncurses
ncurses from 6.4.20231021 enables "NCUSES_OPAQUE_MENU", which hides
some parts of the code away from the application. This causes
petitboot to fail building in a few spots where it tries to directly
access the name of the menu item.

Since the ncurses package was bumped to version 6.5.20250517 in
commit [0], petitboot is failing to build with error:

    ui/ncurses/nc-menu.h: In function ‘pmenu_dump_item’:
    ui/ncurses/nc-menu.h:129:47: error: invalid use of incomplete typedef ‘ITEM’ {aka ‘const struct tagITEM’}
      129 |         pb_debug("%p %s\n", item, (item ? item->name.str : "(null)"));

This commit fixes the issue by adding package patches.
These two patches were submitted to the petitboot mailing list by
Nicholas Piggin [1] resolve this issue using accessors, or by making a new item.

This has been raised upstream but not yet applied [2].

[0]: https://gitlab.com/buildroot.org/buildroot/-/commit/e7c091f11398bdb6498adaa344d0934d97b4ef39
[1]: https://lists.ozlabs.org/pipermail/petitboot/2024-February/thread.html#1545
[2]: https://github.com/open-power/petitboot/issues/106

Signed-off-by: Nathaniel Roach <nroach44@nroach44.id.au>
Fixes: https://autobuild.buildroot.org/results/bf1/bf173e1e0f8ad3f40534aaa358dc40993cd14ac4/build-end.log
Reviewed-by: Laurent Vivier <laurent@vivier.eu>
[Julien:
 - add Signed-off-by in patches
 - add info in commit log
]
Signed-off-by: Julien Olivain <ju.o@free.fr>