package/mosquitto: bump to version 2.0.22
Drop the patch that was fixed in the v2.0.22, see the closed issues and PRs:
https://github.com/eclipse-mosquitto/mosquitto/issues/3246
https://github.com/eclipse-mosquitto/mosquitto/pull/3256
https://github.com/eclipse-mosquitto/mosquitto/pull/3227
https://github.com/eclipse-mosquitto/mosquitto/pull/3252

However, a new issue was reported by GasparQ, when WITH_TLS option is off.
https://github.com/eclipse-mosquitto/mosquitto/issues/3318

Fortunately, besmarsh had pushed a PR that fixed this issue.
https://github.com/eclipse-mosquitto/mosquitto/pull/3321

The patch file was quoted in this commit.

For more details of v2.0.22, see the changelog:
https://github.com/eclipse-mosquitto/mosquitto/blob/v2.0.22/ChangeLog.txt

Signed-off-by: Scott Fan <fancp2007@gmail.com>
Signed-off-by: Julien Olivain <ju.o@free.fr>