package/libspdm: bump version to 3.8.0
For change log since 3.5.0, see:
https://github.com/DMTF/libspdm/releases

Following the upgrade of libopenssl to version 3.5.0 [1], the build
process for libspdm encountered issues due to the new openssl API.
It requires some modifications in the asn1.h file of libspdm.
This build error was addressed by a specific commit [2] in libspdm.

As Buildroot 2025.02 isn't impacted, this commit bumps libspdm to the
latest libspdm version to fix the build error.

This commit also updates the license hash, after a year update in:
https://github.com/DMTF/libspdm/commit/e8a35c8d0dfe5edb2bbafaa07ebf74b097824364

[1]: https://gitlab.com/buildroot.org/buildroot/-/commit/9868ca6ec98afa28a94aea50cc940db7df860396
[2]: https://github.com/DMTF/libspdm/commit/88797f83daf6654d6381e3539abe1670763f2ac1

Fixes:
https://autobuild.buildroot.org/results/87b8805975ef132a58c40e622e2e023199f6ee2a/

Signed-off-by: Thomas Bonnefille <thomas.bonnefille@bootlin.com>
[Julien:
 - add link to change log
 - fix LICENSE.md hash
]
Signed-off-by: Julien Olivain <ju.o@free.fr>