package/bash: fix build with gcc 15
Commit 6d09b25d08fe (package/bash: fix build with host GCC 15) did
apply a workaround for when the host gcc is gcc-15 (or later), but
missed the fact that the target compiler also exhibits build issues.

For the same reasons as explained in 6d09b25d08fe, don't try and fix
those issues, but just paper over the problem by using an older C
standard.

Fixes:

  https://autobuild.buildroot.net/results/1ee74d546182695d3411358f028fcce8ccbb74c3/

Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>
Cc: Adam Duskett <adam.duskett@amarulasolutions.com>
Cc: Florian Larysch <fl@n621.de>
Cc: Julien Olivain <ju.o@free.fr>
Reviewed-by: Florian Larysch <fl@n621.de>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>